drm/rockchip: vop: Add support mipi TE mode
For software TE mode, we register a gpio IRQ to respond to the TE
signal from the panel. If the falling edge is detected via gpio, a
new frame will be sent to the panel for display by software driver.

For hardware TE mode, we set the specific pin to TE function iomux.
If the TE signal is detected, a new frame will be sent to the RX
device by hardware automatically.

Change-Id: Ic594e5eb4b267a52441fb2720c788aba4d8ffb72
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>