ARM: configs: rockchip: Migrate configs for Bifrost and Valhall GPUs to SoC-level config files
This is in coordination with the modification of commit
"MALI: rockchip: Add separate src directory for Valhall driver from DDK g28p0-00eac0".
The configs related to Bifrost and Valhall GPUs are designed to be configured
in the SoC-level config files.
Therefore, for some SoCs, the kernel compilation command line needs to be adjusted.

Change-Id: Iadb02e7f7cf6cde7069b16675475b81925d37e13
Signed-off-by: Li Huang <putin.li@rock-chips.com>
Signed-off-by: Zhen Chen <chenzhen@rock-chips.com>